feat: Modularize handlers
